Refactor test data loading logic and add recursive search for test data files
